remove int cast

+++ Assets/Scripts/tutorial/TutorialStyleGiver.cs
@@ -5,7 +5,7 @@ public class TutorialStyleGiver : MonoBehaviour, IDamageable
     public PointManager pointManager;
     public void ApplyDamage(Weapon weapon, float amount)
     {
-        pointManager.AddPoints((int)(amount * 500));
+        pointManager.AddPoints(amount * 500);
     }
 
 }
